+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Rename function `plot_kernel_jacobians` to `jacobian_plot` in `autograd` module.
 - Fix module hashing when a kernel argument contained a struct array ([GH-287](https://github.com/NVIDIA/warp/issues/287)).
 - Add support for fp64 `atomic_add`, `atomic_max`, and `atomic_min` ([GH-284](https://github.com/NVIDIA/warp/issues/284)).
+- Add `wp.mod()` for vector types ([GH-282](https://github.com/NVIDIA/warp/issues/282)).
 
 ## [1.3.1] - 2024-07-27